feat: implement FlowAiAgentBase class to streamline AI agent functionality and enhance code reuse across FlowAiAgent and FlowAiAgentRouter
