+++ packages/botonic-plugin-flow-builder/src/content-fields/flow-ai-agent-router.tsx
@@ -1,31 +1,19 @@
 import {
-  type AgenticOutputMessage,
   AiAgentType,
   type BotContext,
-  type GuardrailRule,
-  type HubtypeAssistantMessage,
   type InferenceResponse,
-  OutputMessageType,
-  VerbosityLevel,
 } from '@botonic/core'
 import type { FlowBuilderApi } from '../api'
 import {
   type FlowBuilderContentMessage,
   FlowBuilderContentSchema,
-  FlowBuilderOutputMessageType,
 } from '../structured-output/flow-builder-content'
 import { getCommonFlowContentEventArgsForContentId } from '../tracking'
-import { HubtypeAssistantContent } from '../utils/ai-agent'
 import { getFlowBuilderPlugin } from '../utils/get-flow-builder-plugin'
-import { ContentFieldsBase } from './content-fields-base'
 import { FlowAiAgent } from './flow-ai-agent'
-import type {
-  HtAiAgentNode,
-  HtAiAgentRouterNode,
-  HtInputGuardrailRule,
-  HtNodeWithContent,
-} from './hubtype-fields'
-import { FlowCarousel, type FlowContent, FlowText } from './index'
+import { FlowAiAgentBase } from './flow-ai-agent-base'
+import type { HtAiAgentNode, HtAiAgentRouterNode } from './hubtype-fields'
+import type { FlowContent } from './index'
 
 export interface AiAgentWithNameAndDescription {
   agent: FlowAiAgent
@@ -33,17 +21,8 @@ export interface AiAgentWithNameAndDescription {
   name: string
 }
 
-export class FlowAiAgentRouter extends ContentFieldsBase {
-  public name: string = ''
-  public instructions: string = ''
-  public model: string = ''
-  public verbosity: VerbosityLevel = VerbosityLevel.Medium
+export class FlowAiAgentRouter extends FlowAiAgentBase {
   public agents: AiAgentWithNameAndDescription[] = []
-  public inputGuardrailRules: HtInputGuardrailRule[] = []
-
-  public aiAgentResponse?: InferenceResponse<FlowBuilderContentMessage>
-  public messages: AgenticOutputMessage<FlowBuilderContentMessage>[] = []
-  public jsxElements: JSX.Element[] = []
 
   static fromHubtypeCMS(
     component: HtAiAgentRouterNode,
@@ -93,36 +72,10 @@ export class FlowAiAgentRouter extends ContentFieldsBase {
     return name?.trim().toLowerCase().replace(/ /g, '_') || `ai_agent_${index}`
   }
 
-  async resolveAIAgentResponse(
-    botContext: BotContext,
-    previousContents?: FlowContent[]
-  ): Promise<InferenceResponse<FlowBuilderContentMessage> | undefined> {
-    const aiAgentResponse = await this.getAIAgentResponse(
-      botContext,
-      previousContents
-    )
-
-    if (aiAgentResponse) {
-      this.aiAgentResponse = aiAgentResponse
-      await this.trackAiAgentResponse(botContext)
-      this.messages = aiAgentResponse.messages
-    }
-
-    return aiAgentResponse
-  }
-
   async getAIAgentResponse(
     botContext: BotContext,
     previousContents?: FlowContent[]
   ): Promise<InferenceResponse<FlowBuilderContentMessage> | undefined> {
-    const previousHubtypeContents: HubtypeAssistantMessage[] =
-      previousContents?.map(content => {
-        return {
-          role: 'assistant',
-          content: HubtypeAssistantContent.adapt(content),
-        }
-      }) || []
-
     const flowBuilderPlugin = getFlowBuilderPlugin(botContext.plugins)
     const aiAgentResponse = await flowBuilderPlugin.getAiAgentResponse?.(
       botContext,
@@ -149,7 +102,8 @@ export class FlowAiAgentRouter extends ContentFieldsBase {
           this.inputGuardrailRules
         ),
         outputMessagesSchemas: [FlowBuilderContentSchema],
-        previousHubtypeMessages: previousHubtypeContents,
+        previousHubtypeMessages:
+          this.getPreviousHubtypeContents(previousContents),
       }
     )
 
@@ -158,10 +112,6 @@ export class FlowAiAgentRouter extends ContentFieldsBase {
     })
 
     return aiAgentResponse
-  }
-
-  async trackFlow(_botContext: BotContext): Promise<void> {
-    return
   }
 
   async trackAiAgentResponse(botContext: BotContext) {
@@ -193,80 +143,5 @@ export class FlowAiAgentRouter extends ContentFieldsBase {
       eventArgs,
     })
     // await trackEvent(botContext, action, eventArgs)
-  }
-
-  async getFlowContentsByContentId(
-    botContext: BotContext,
-    contentId: string
-  ): Promise<FlowContent[]> {
-    const flowBuilderPlugin = getFlowBuilderPlugin(botContext.plugins)
-    const cmsApi = flowBuilderPlugin.cmsApi
-    const node = cmsApi.getNodeByContentID(contentId)
-    const flowContents = await flowBuilderPlugin.getContentsByNode(
-      node as HtNodeWithContent
-    )
-
-    return flowContents
-  }
-
-  async messagesToBotonicJSXElements(botContext: BotContext): Promise<void> {
-    for (const message of this.messages) {
-      if (
-        message.type === OutputMessageType.Text ||
-        message.type === OutputMessageType.TextWithButtons ||
-        message.type === OutputMessageType.BotExecutor
-      ) {
-        this.jsxElements.push(FlowText.fromAIAgent(this.id, message))
-      }
-
-      if (message.type === OutputMessageType.Carousel) {
-        this.jsxElements.push(
-          FlowCarousel.fromAIAgent(this.id, message, botContext)
-        )
-      }
-
-      if (message.type === FlowBuilderOutputMessageType.FlowBuilderContent) {
-        const flowContents = await this.getFlowContentsByContentId(
-          botContext,
-          message.contentId
-        )
-        for (const content of flowContents) {
-          await content.processContent(botContext)
-          this.jsxElements.push(content.toBotonic(botContext))
-        }
-      }
-    }
-    return
-  }
-
-  private getActiveInputGuardrailRules(
-    inputGuardrailRules: HtInputGuardrailRule[]
-  ): GuardrailRule[] {
-    return (
-      inputGuardrailRules
-        ?.filter(rule => rule.is_active)
-        ?.map(rule => ({
-          name: rule.name,
-          description: rule.description,
-        })) || []
-    )
-  }
-
-  async processContent(
-    botContext: BotContext,
-    previousContents?: FlowContent[]
-  ): Promise<void> {
-    if (this.messages.length === 0) {
-      await this.resolveAIAgentResponse(botContext, previousContents)
-    }
-    if (this.jsxElements.length === 0) {
-      await this.filterContent(botContext, this)
-      await this.messagesToBotonicJSXElements(botContext)
-    }
-    return
-  }
-
-  toBotonic(): JSX.Element {
-    return <>{this.jsxElements}</>
   }
 }
